GH-2997: Move ARP's internal NTriples reader to tests, and rename
