+++ jena-core/src/test/java/org/apache/jena/rdfxml/xmlinput0/NTripleARP0.java
@@ -22,7 +22,7 @@
  * Created on July 13, 2001, 10:06 PM
  */
 
-package org.apache.jena.rdfxml.xmlinput1;
+package org.apache.jena.rdfxml.xmlinput0;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -83,12 +83,13 @@ import org.xml.sax.SAXParseException;
  * </dt><dd>
  * Ignores numbered error/warning conditions.
  * </dl>
+ * @deprecated Legacy. To be removed.
  */
-@SuppressWarnings("deprecation")
-public class NTriple implements ARPErrorNumbers {
+@Deprecated(forRemoval = true)
+public class NTripleARP0 implements ARPErrorNumbers {
 
 	private static StringBuilder line = new StringBuilder();
-	private static ARP arp;
+    private static ARP0 arp;
 	private static String xmlBase = null;
 	private static boolean numbers = false;
 	/** Starts an RDF/XML to NTriple converter.
@@ -112,7 +113,7 @@ public class NTriple implements ARPErrorNumbers {
 		andMeToo = ap;
 		//SH sh = new SH();
 		int i;
-		arp = new ARP();
+		arp = new ARP0();
 		ARPHandlers handlers = arp.getHandlers();
 		handlers.setStatementHandler(getSH(true));
 //        arp.getOptions().setEmbedding(true);
@@ -197,7 +198,7 @@ public class NTriple implements ARPErrorNumbers {
 	static void usage() {
 		System.err.println(
 			"java <class-path> "
-				+ NTriple.class.getName()
+				+ NTripleARP0.class.getName()
 				+ " ( [ -[xstfurR]][ -b xmlBase -[eiw] NNN[,NNN...] ] [ file ] [ url ] )... ");
 		System.err.println(
 			"    All options, files and URLs can be intemingled in any order.");
