+++ jena-core/src/test/java/org/apache/jena/rdfxml/xmlinput1/NTripleTestSuite.java
@@ -226,6 +226,7 @@ class NTripleTestSuite extends WGTestSuite {
 		return loadRDFx(in, eh, base, true, 0);
 	}
 
+    @SuppressWarnings("removal")
     static Model loadRDFx(
 		InputSupplier in,
 		RDFErrorHandler eh,
@@ -257,7 +258,7 @@ class NTripleTestSuite extends WGTestSuite {
 			System.setIn(in.open());
 			System.setOut(out);
 			try {
-				NTriple.mainEh(new String[] { "-b", base, "-s" }, th, th);
+				NTripleARP1.mainEh(new String[] { "-b", base, "-s" }, th, th);
 			} catch (SimulatedException e) {
 				if (wantModel)
 					throw e;
