+++ jena-core/src/test/java/org/apache/jena/rdfxml/xmlinput1/TestsTainting.java
@@ -183,6 +183,7 @@ public class TestsTainting extends TestCase implements ErrorHandler,
 		OntDocumentManager.getInstance().reset(true);
 	}
 
+    @SuppressWarnings("removal")
     @Override
     public void runTest() throws IOException {
         ByteArrayOutputStream goodBytes = new ByteArrayOutputStream();
@@ -192,7 +193,7 @@ public class TestsTainting extends TestCase implements ErrorHandler,
         try (PrintStream out = new PrintStream(goodBytes); PrintStream err = new PrintStream(badBytes); ) {
             System.setOut(out);
             System.setErr(err);
-            NTriple.mainEh(new String[]{"-e","102,136,105,103,108,107,116,106,004,131",
+            NTripleARP1.mainEh(new String[]{"-e","102,136,105,103,108,107,116,106,004,131",
                 "-E","-b",base,fileName},this,null);
         }
         finally {
